requirements.txt
anthropic>=0.8.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn
pyodbc>=4.0.0
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.0
tqdm>=4.64.0

# Azure dependencies para soporte completo
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
